+++ src/nco/nco_var_utl.h
@@ -1,4 +1,4 @@
-/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_var_utl.h,v 1.50 2009-05-01 22:31:24 zender Exp $ */
+/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_var_utl.h,v 1.51 2009-05-02 22:22:30 zender Exp $ */
 
 /* Purpose: Variable utilities */
 
@@ -48,7 +48,7 @@ extern "C" {
    const int out_id, /* I [id] netCDF output file ID */
    const int rec_dmn_id, /* I [id] Input file record dimension ID  */
    const char * const var_nm, /* I [sng] Input variable name */
-   const int *cnk_sz, /* I [nbr] Chunk sizes */
+   const size_t *cnk_sz, /* I [nbr] Chunk sizes */
    const int dfl_lvl); /* I [enm] Deflate level [0..9] */
 
   int /* O [id] Output file variable ID */
@@ -59,7 +59,7 @@ extern "C" {
    const char * const var_nm, /* I [sng] Input variable name */
    CST_X_PTR_CST_PTR_CST_Y(lmt_all_sct,lmt_all_lst), /* I [sct] Hyperslab limits */
    const int lmt_all_lst_nbr, /* I [nbr] Number of hyperslab limits */
-   const int *cnk_sz, /* I [nbr] Chunk sizes */
+   const size_t *cnk_sz, /* I [nbr] Chunk sizes */
    const int dfl_lvl); /* I [enm] Deflate level [0..9] */
 
   void
@@ -102,7 +102,7 @@ extern "C" {
    const int nbr_dmn_ncl, /* I [nbr] Number of dimensions in list */
    const int nco_pck_map, /* I [enm] Packing map */
    const int nco_pck_plc, /* I [enm] Packing policy */
-   const int *cnk_sz, /* I [nbr] Chunk sizes */
+   const size_t *cnk_sz, /* I [nbr] Chunk sizes */
    const int dfl_lvl); /* I [enm] Deflate level [0..9] */
   
   var_sct * /* O [sct] Copy of input variable */
